Fix - github service moved to utils
