--- libs/github/index.ts
@@ -1,1 +1,0 @@
-export * from './src';
