--- libs/github/ng-package.prod.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/libs/github",
-  "lib": {
-    "entryFile": "src/index.ts"
-  }
-}
