--- libs/github/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@codelab/github",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0"
-  }
-}
