--- libs/github/tslint.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tslint.json",
-  "rules": {
-    "directive-selector": [true, "attribute", "github", "camelCase"],
-    "component-selector": [true, "element", "github", "kebab-case"]
-  }
-}
