+++ libs/utils/src/lib/github-PR-service/github.service.ts
@@ -58,6 +58,8 @@ export class GithubService {
 
   async createClosedIssue(message, reason) {
 
+    console.log(this.githubAuth);
+
     if (!this.githubAuth.credential) {
       await this.login();
     }
