+++ libs/utils/src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './button-with-menu';
 export * from './full-screen-mode';
+export * from './github-PR-service/index';
